refactor(#178): share the replay engine call with dispatch
Splits the wire-input/engine-output conversion and the simulation
call out of the oRPC handler into their own module, with no behaviour
change, so the in-process dispatch path runs the same logic the
replay endpoint does.
